+++ archive_download.py
@@ -1,6 +1,6 @@
 import os
-import math
 import PySimpleGUI as sg
+import common_functions as common
 from urllib.parse import quote
 from urllib.request import urlopen
 from http.client import HTTPResponse
@@ -25,7 +25,7 @@ def download_files(platform_id: str, filenames: list, output_folder: str, cb_pro
       output_stream.flush()
       position += len(bytes_read)
       cb_progressbar.update(position, file_length)
-      cb_text_status.update(f"{length_to_unit_string(position)} / {length_to_unit_string(file_length)}")
+      cb_text_status.update(f"{common.length_to_unit_string(position)} / {common.length_to_unit_string(file_length)}")
       if len(bytes_read) < buffer_length: break
     
     file_request.close()
@@ -37,22 +37,3 @@ def download_files(platform_id: str, filenames: list, output_folder: str, cb_pro
     cb_text_status.update("0Kb / 0Kb")
     download(filename)
   cb_function()
-
-
-def length_to_unit_string(length: float) -> str:
-  unit = "b"
-  one_kilobytes = math.pow(1024, 1)
-  one_megabytes = math.pow(1024, 2)
-  one_gigabytes = math.pow(1024, 3)
-
-  if length > one_kilobytes and length < one_megabytes:
-    length /= one_kilobytes
-    unit = "Kb"
-  if length > one_megabytes and length < one_gigabytes:
-    length /= one_megabytes
-    unit = "Mb"
-  if length > one_gigabytes:
-    length /= one_gigabytes
-    unit = "Gb"
-  
-  return "{:.1f}{}".format(length, unit)