+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * A way less I/O network operations since I discovered a "JSON" manner to handle data scrapped from [Internet Archive](https://archive.org).
 * Changed games combo view with a listview to be able to download for than one file at once.
 * Added cursors
+* Added file's details window
 * Options zone added but not yet ready (Unzip & more to come)
 
 ## v1.1